+++ src/Interceptor/WorkflowClient/UpdateRef.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Temporal\Workflow\Update;
+namespace Temporal\Interceptor\WorkflowClient;
 
 use Temporal\Internal\Marshaller\Meta\Marshal;
 use Temporal\Workflow\WorkflowExecution;
