+++ src/Interceptor/WorkflowClientCallsInterceptor.php
@@ -19,10 +19,10 @@ use Temporal\Interceptor\WorkflowClient\QueryInput;
 use Temporal\Interceptor\WorkflowClient\SignalInput;
 use Temporal\Interceptor\WorkflowClient\SignalWithStartInput;
 use Temporal\Interceptor\WorkflowClient\StartInput;
+use Temporal\Interceptor\WorkflowClient\StartUpdateOutput;
 use Temporal\Interceptor\WorkflowClient\TerminateInput;
 use Temporal\Interceptor\WorkflowClient\UpdateInput;
 use Temporal\Internal\Interceptor\Interceptor;
-use Temporal\Workflow\Update\StartUpdateOutput;
 use Temporal\Workflow\WorkflowExecution;
 
 /**
@@ -51,7 +51,7 @@ interface WorkflowClientCallsInterceptor extends Interceptor
 
     /**
      * @param UpdateInput $input
-     * @param callable(UpdateInput): void $next
+     * @param callable(UpdateInput): StartUpdateOutput $next
      *
      * @return StartUpdateOutput
      */
